backend/requirements.txt
fastapi==0.133.1
uvicorn[standard]==0.41.0
gunicorn==25.1.0
duckdb==1.4.4
pandas==2.3.3
numpy==2.2.6
pandas_ta==0.4.71b0
python-dotenv==1.2.1
pydantic==2.12.5
# Optional: install TA-Lib for faster WMA/ADX/Williams%R (requires system lib)
# TA-Lib>=0.4.28
